Update CSS styles for responsive design
The hard-coded height values in certain CSS classes were removed and changed to 'auto' to accommodate dynamic content lengths. This adjustment makes the design more responsive and fits better with different content sizes. The min-height property of the .cardEventLateralMovil class was also removed for a more flexible layout on mobile view.